Compulsory Question for all groups
Write a C program to simulate a guessing game that awards a user points based on their guess. The 
user is given 2 attempts to guess the number 79. If the user guesses the number during the first 
attempt should return the points earned p (p=10). If the user guesses the number during the second 
attempt, the program should return the points earned 0.8p (80% of p). If after two attempts the user
fails to guess correctly the program should use the following equation to calculate the points to be 
returned: ((80-(79-g)
 2)/100) p. Where g is the user’s guess. [15]